netqmail-1.05-tls-20070408.patch

+++ ssl_timeoutio.h
@@ -8,14 +8,14 @@
 # error "Need OpenSSL version at least 0.9.6"
 #endif
 
-int ssl_timeoutconn(long t, int rfd, int wfd, SSL *ssl);
-int ssl_timeoutaccept(long t, int rfd, int wfd, SSL *ssl);
-int ssl_timeoutrehandshake(long t, int rfd, int wfd, SSL *ssl);
+int ssl_timeoutconn(int t, int rfd, int wfd, SSL *ssl);
+int ssl_timeoutaccept(int t, int rfd, int wfd, SSL *ssl);
+int ssl_timeoutrehandshake(int t, int rfd, int wfd, SSL *ssl);
 
-int ssl_timeoutread(long t, int rfd, int wfd, SSL *ssl, char *buf, int len);
-int ssl_timeoutwrite(long t, int rfd, int wfd, SSL *ssl, char *buf, int len);
+int ssl_timeoutread(int t, int rfd, int wfd, SSL *ssl, char *buf, int len);
+int ssl_timeoutwrite(int t, int rfd, int wfd, SSL *ssl, char *buf, int len);
 
 int ssl_timeoutio(
-  int (*fun)(), long t, int rfd, int wfd, SSL *ssl, char *buf, int len);
+  int (*fun)(), int t, int rfd, int wfd, SSL *ssl, char *buf, int len);
 
 #endif
